댓글 작성, 수정, 삭제 기능 구현
성공, 예외 응답은 임시로 구현

+++ src/main/java/com/example/troubleshooter/service/CommentService.java
@@ -6,7 +6,9 @@ import com.example.troubleshooter.entity.Post;
 import com.example.troubleshooter.repository.CommentRepository;
 import com.example.troubleshooter.repository.PostRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 @RequiredArgsConstructor
 @Service
@@ -16,21 +18,31 @@ public class CommentService {
 
     private final PostRepository postRepository;
 
-    public void writeComment(Long postId, CommentRequestDto commentRequestDto) {
+    public void writeComment(Long postId, CommentRequestDto commentRequestDto, Long userId) {
         Post post = postRepository.findById(postId)
-                .orElseThrow(() -> new IllegalArgumentException("게시글을 찾을 수 없습니다."));
-        post.addComment(new Comment(commentRequestDto));
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "게시글을 찾을 수 없습니다."));
+        post.addComment(new Comment(commentRequestDto, userId));
     }
 
-    public void editComment(Long commentId, CommentRequestDto commentRequestDto) {
+    public void editComment(Long commentId, CommentRequestDto commentRequestDto, Long userId) {
         Comment comment = commentRepository.findById(commentId)
-                .orElseThrow(() -> new IllegalArgumentException("댓글을 찾을 수 없습니다."));
-        comment.setContent(commentRequestDto.getContent());
-        commentRepository.save(comment);
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "댓글을 찾을 수 없습니다."));
+        if (comment.isWriter(userId)) {
+            comment.setContent(commentRequestDto.getContent());
+            commentRepository.save(comment);
+        } else {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "댓글 작성자만 수정할 수 있습니다.");
+        }
     }
 
-    public void deleteComment(Long commentId) {
-        commentRepository.deleteById(commentId);
+    public void deleteComment(Long commentId, Long userId) {
+        Comment comment = commentRepository.findById(commentId)
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "댓글을 찾을 수 없습니다."));
+        if (comment.isWriter(userId)) {
+            commentRepository.deleteById(commentId);
+        } else {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "댓글 작성자만 삭제할 수 있습니다.");
+        }
     }
 
 }
